+++ hephaistos-data/lua/Filters/RoomPresentation.lua
@@ -1,0 +1,5 @@
+-- fix biome map (world map shown between)
+Hephaistos.Attach[RunBiomePresentation] = function(params)
+  return params.DestinationId == ScreenAnchors.RunDepthDisplayAnchor
+		and params.OffsetX and params.OffsetY
+end