+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+-   Fix boon info screen.
 -   Fix quest log (fates) scroll up/down buttons.
 
 ## [v0.7.0] - 2021-08-19