+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Fix boon info screen.
+-   Fix mid-run trait overlay (B button header).
 -   Fix quest log (fates) scroll up/down buttons.
 
 ## [v0.7.0] - 2021-08-19