+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Fix boon info screen.
 -   Fix mid-run trait overlay (B button header).
+-   Fix world map shown between regions.
 -   Fix quest log (fates) scroll up/down buttons.
 
 ## [v0.7.0] - 2021-08-19